+++ lib/bundler/cli.rb
@@ -672,10 +672,13 @@ module Bundler
     end
 
     desc "clean", "Cleans up unused gems in your bundler directory"
+    method_option "dry-run", :type => :boolean, :default => false, :banner =>
+      "only print out changes, do not actually clean gems"
     method_option "force", :type => :boolean, :default => false, :banner =>
       "forces clean even if --path is not set"
     def clean
       if Bundler.settings[:path] || options[:force]
+        Bundler.settings[:dry_run] = options[:"dry-run"]
         Bundler.load.clean
       else
         Bundler.ui.error "Can only use bundle clean when --path is set or --force is set"